all transactions count at top addresses page
