+++ apps/block_scout_web/lib/block_scout_web/templates/robots/sitemap.xml.eex
@@ -11,7 +11,7 @@
     </url>
   <% end %>
 
-  <% addresses = Chain.list_top_addresses(params) %>
+  <% addresses = Address.list_top_addresses(params) %>
   <%= for {address, _} <- addresses do %>
     <url>
       <loc><%= host %>/address/<%= to_string(address) %></loc>
