+++ apps/block_scout_web/test/block_scout_web/controllers/api/v2/address_controller_test.exs
@@ -1708,7 +1708,7 @@ defmodule BlockScoutWeb.API.V2.AddressControllerTest do
     end
 
     test "check nil", %{conn: conn} do
-      address = insert(:address, nonce: 1, fetched_coin_balance: 1)
+      address = insert(:address, transactions_count: 2, fetched_coin_balance: 1)
 
       request = get(conn, "/api/v2/addresses")
 
@@ -2477,7 +2477,7 @@ defmodule BlockScoutWeb.API.V2.AddressControllerTest do
 
   defp compare_item(%Address{} = address, json) do
     assert Address.checksum(address.hash) == json["hash"]
-    assert to_string(address.nonce + 1) == json["tx_count"]
+    assert to_string(address.transactions_count) == json["tx_count"]
   end
 
   defp compare_item(%Transaction{} = transaction, json) do
