+++ apps/explorer/test/explorer/chain_test.exs
@@ -2058,76 +2058,6 @@ defmodule Explorer.ChainTest do
     end
   end
 
-  describe "list_top_addresses/0" do
-    test "without addresses with balance > 0" do
-      insert(:address, fetched_coin_balance: 0)
-      assert [] = Chain.list_top_addresses()
-    end
-
-    test "with top addresses in order" do
-      address_hashes =
-        4..1
-        |> Enum.map(&insert(:address, fetched_coin_balance: &1))
-        |> Enum.map(& &1.hash)
-
-      assert address_hashes ==
-               Chain.list_top_addresses()
-               |> Enum.map(fn {address, _transaction_count} -> address end)
-               |> Enum.map(& &1.hash)
-    end
-
-    # flaky test
-    # test "with top addresses in order with matching value" do
-    #   test_hashes =
-    #     4..0
-    #     |> Enum.map(&Explorer.Chain.Hash.cast(Explorer.Chain.Hash.Address, &1))
-    #     |> Enum.map(&elem(&1, 1))
-
-    #   tail =
-    #     4..1
-    #     |> Enum.map(&insert(:address, fetched_coin_balance: &1, hash: Enum.fetch!(test_hashes, &1 - 1)))
-    #     |> Enum.map(& &1.hash)
-
-    #   first_result_hash =
-    #     :address
-    #     |> insert(fetched_coin_balance: 4, hash: Enum.fetch!(test_hashes, 4))
-    #     |> Map.fetch!(:hash)
-
-    #   assert [first_result_hash | tail] ==
-    #            Chain.list_top_addresses()
-    #            |> Enum.map(fn {address, _transaction_count} -> address end)
-    #            |> Enum.map(& &1.hash)
-    # end
-
-    # flaky test
-    # test "paginates addresses" do
-    #   test_hashes =
-    #     4..0
-    #     |> Enum.map(&Explorer.Chain.Hash.cast(Explorer.Chain.Hash.Address, &1))
-    #     |> Enum.map(&elem(&1, 1))
-
-    #   result =
-    #     4..1
-    #     |> Enum.map(&insert(:address, fetched_coin_balance: &1, hash: Enum.fetch!(test_hashes, &1 - 1)))
-    #     |> Enum.map(& &1.hash)
-
-    #   options = [paging_options: %PagingOptions{page_size: 1}]
-
-    #   [{top_address, _}] = Chain.list_top_addresses(options)
-    #   assert top_address.hash == List.first(result)
-
-    #   tail_options = [
-    #     paging_options: %PagingOptions{key: {top_address.fetched_coin_balance.value, top_address.hash}, page_size: 3}
-    #   ]
-
-    #   tail_result = tail_options |> Chain.list_top_addresses() |> Enum.map(fn {address, _} -> address.hash end)
-
-    #   [_ | expected_tail] = result
-
-    #   assert tail_result == expected_tail
-    # end
-  end
-
   describe "stream_blocks_without_rewards/2" do
     test "includes consensus blocks" do
       %Block{hash: consensus_hash} = insert(:block, consensus: true)
